ai-agent/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
web3==6.11.1
schedule==1.2.0 